pagination impelemented

+++ src/main/java/com/thepetclub/ProductService/request/ProductUpdateRequest.java
@@ -1,16 +1,13 @@
 package com.thepetclub.ProductService.request;
 
-import com.thepetclub.ProductService.model.Category;
 import lombok.Data;
-import org.bson.types.ObjectId;
 
-import java.math.BigDecimal;
 
 @Data
 public class ProductUpdateRequest {
     private String  id;
     private String name;
-    private BigDecimal price;
+    private Double price;
     private int inventory;
     private String description;
     private String categoryName;
